Add more message types

+++ api/src/main/java/com/illtamer/infinite/bot/api/entity/transfer/XML.java
@@ -16,6 +16,6 @@ public class XML implements TransferEntity {
     private String data;
 
     @Nullable
-    private Integer resid;
+    private Short resid;
 
 }
